requirements.txt
fastapi
uvicorn
aiosqlite
python-dotenv
pyjwt